// projects/documentation-site/src/app/generated-doc-routes.ts
export const GENERATED_DOC_ROUTES: Routes = [
  {
    path: 'composables',
    title: 'Composables',
    children: [
      {
        path: 'browser',
        title: 'Browser',
        children: [
          {
            path: 'use-document-visibility',
            component: UseDocumentVisibilityPageComponent,
            title: 'useDocumentVisibility',
          },
          {
            path: 'use-mouse-position',
            component: UseMousePositionPageComponent,
            title: 'useMousePosition',
          },
          {
            path: 'use-window-size',
            component: UseWindowSizePageComponent,
            title: 'useWindowSize',
          },
        ],
      },
      {
        path: 'route',
        title: 'Route',
        children: [
          {
            path: 'use-parameters',
            component: UseParametersPageComponent,
            title: 'useParameters',
          },
          {
            path: 'use-query-parameters',
            component: UseQueryParametersPageComponent,
            title: 'useQueryParameters',
          },
          {
            path: 'use-route-data',
            component: UseRouteDataPageComponent,
            title: 'useRouteData',
          },
          {
            path: 'use-route-fragment',
            component: UseRouteFragmentPageComponent,
            title: 'useRouteFragment',
          },
          {
            path: 'use-route-params',
            component: UseRouteParamsPageComponent,
            title: 'useRouteParams',
          },
          {
            path: 'use-route-query-param',
            component: UseRouteQueryParamPageComponent,
            title: 'useRouteQueryParam',
          },
        ],
      },
      {
        path: 'use-debounced-signal',
        component: UseDebouncedSignalPageComponent,
        title: 'useDebouncedSignal',
      },
      {
        path: 'use-previous-signal',
        component: UsePreviousSignalPageComponent,
        title: 'usePreviousSignal',
      },
      {
        path: 'use-throttled-signal',
        component: UseThrottledSignalPageComponent,
        title: 'useThrottledSignal',
      },
    ],
  },
  {
    path: 'effects',
    title: 'Effects',
    children: [
      {
        path: 'log-changes',
        component: LogChangesPageComponent,
        title: 'logChanges',
      },
      {
        path: 'sync-local-storage',
        component: SyncLocalStoragePageComponent,
        title: 'syncLocalStorage',
      },
      {
        path: 'sync-query-params',
        component: SyncQueryParamsPageComponent,
        title: 'syncQueryParams',
      },
    ],
  },
  {
    path: 'utils',
    title: 'Utils',
    children: [
      {
        path: 'create-shared-composable',
        component: CreateSharedComposablePageComponent,
        title: 'createSharedComposable',
      },
      {
        path: 'create-singleton-composable',
        component: CreateSingletonComposablePageComponent,
        title: 'createSingletonComposable',
      },
    ],
  },
];